requirements.txt
numpy==1.22.1
opencv-python==4.5.5.62

git+https://github.com/huggingface/transformers.git
git+https://github.com/Ulbwaa/YandexImagesParser.git
sentencepiece
natasha
colour
deep-translator
torch
python-telegram-bot
nltk
